small edits
