+++ musasampo-frontend/src/App.js
@@ -1,4 +1,4 @@
-import { Switch, Route, Redirect, Link } from 'react-router-dom';
+import { Switch, Route } from 'react-router-dom';
 
 import HomePage from './pages/home-page/home-page.component';
 import ShopPage from './pages/shop-page/shop-page.component';
@@ -14,13 +14,12 @@ import BandPage from './components/band/band.component';
 import AlbumComponent from './components/album/album.component';
 import Header from './components/header/header.component';
 import LoginPage from './pages/sign-in-and-sign-up/sign-in-and-sign-up.component';
-import Profile from './pages/profile/profile.component';
+import ProfilePage from './pages/profile-page/profile-page.component';
 
-import { stringify } from 'querystring';
-import ProfileEdit from './pages/profile-edit/profile-edit.component';
-import ProfileChangePassword from './pages/profile-change-password/profile-change-password.component';
-import NewBand from './pages/profile-new-band/profile-new-band.component';
-import NewAlbum from './pages/profile-new-album/profile-new-album.component';
+import ProfileEditPage from './components/profile-edit/profile-edit.component';
+import ProfileChangePasswordPage from './components/profile-change-password/profile-change-password.component';
+import NewBandPage from './components/profile-new-band/profile-new-band.component';
+import NewAlbumPage from './components/profile-new-album/profile-new-album.component';
 
 
 import React, { Component } from "react";
@@ -52,30 +51,30 @@ class App extends Component {
             <div className="App">
                 <header className="App-header">
                 <Header />
+                </header>
+                <body>
                     <Switch>
-                      <Route exact path='/' component={HomePage} />
-                      <PrivateRoute exact path='/chat' component={ChatPage} />
-                      <Route exact path='/login' component={LoginPage} />
-                      <PrivateRoute exact path='/mp3-upload' component={Mp3} />
-                      <PrivateRoute exact path='/shop' component={ShopPage} />
-                      <Route exact path='/search' component={SearchPage}/>
-                      <PrivateRoute exact path='/profile' component={Profile} />
-                      <PrivateRoute exact path='/profile/edit' component={ProfileEdit} />
-                      <PrivateRoute exact path='/profile/password' component={ProfileChangePassword} />
-                      <PrivateRoute exact path='/profile/newband' component={NewBand} />
-                      <PrivateRoute exact path='/profile/newalbum' component={NewAlbum} />
-                      <PrivateRoute exact path='/bands/:bandId' component={BandPage} /> 
-                      <PrivateRoute exact path='/shop/albums/genre/:genre' component={GenrePage} />          
-                      <PrivateRoute exact path='/shop/albums/:albumId' component={AlbumComponent} />
-                      <PrivateRoute exact path='/guitartuner' component={GuitarOverview} />    
-                      <PrivateRoute exact path='/guitartuner/:guitarId' component={GuitarTuner} />  
-                    </Switch>
-                </header> 
-
+                        <Route exact path='/' component={HomePage} />
+                        <PrivateRoute exact path='/chat' component={ChatPage} />
+                        <Route exact path='/login' component={LoginPage} />
+                        <PrivateRoute exact path='/mp3-upload' component={Mp3} />
+                        <PrivateRoute exact path='/shop' component={ShopPage} />
+                        <Route exact path='/search' component={SearchPage}/>
+                        <PrivateRoute exact path='/profile' component={ProfilePage} />
+                        <PrivateRoute exact path='/profile/edit' component={ProfileEditPage} />
+                        <PrivateRoute exact path='/profile/password' component={ProfileChangePasswordPage} />
+                        <PrivateRoute exact path='/profile/newband' component={NewBandPage} />
+                        <PrivateRoute exact path='/profile/newalbum' component={NewAlbumPage} />
+                        <PrivateRoute exact path='/bands/:bandId' component={BandPage} /> 
+                        <PrivateRoute exact path='/shop/albums/genre/:genre' component={GenrePage} />          
+                        <PrivateRoute exact path='/shop/albums/:albumId' component={AlbumComponent} />
+                        <PrivateRoute exact path='/guitartuner' component={GuitarOverview} />    
+                        <PrivateRoute exact path='/guitartuner/:guitarId' component={GuitarTuner} />  
+                    </Switch>  
+                </body>
             </div>
         );
     }
 }
 
-//<p className="App-intro">{this.state.apiResponse}</p>
 export default App;
