+++ musasampo-frontend/src/components/form-input/form-input.styles.scss
@@ -9,8 +9,8 @@ $main-color: black;
 
 .group {
   position: relative;
-  margin: 35px 0;
-
+  margin-bottom: 35px;
+  
   .form-input {
     background: none;
     background-color: white;
