+++ musasampo-frontend/src/components/profile-change-password/profile-change-password.styles.scss
@@ -1,7 +1,7 @@
 .container {
     width: 95%;
 
-    .edit {
+    .sign-up {
         width: 380px;
         display: flex;
         flex-direction: column;
@@ -12,14 +12,10 @@
         margin-top: 40px;
         }
     
-        .buttons1 {
-        display: flex;
-        justify-content: space-between;
-        padding-top: 15px;
-        }
-
-        .form-input-label {
-            color: black;
+        .buttons {
+            display: flex;
+            justify-content: space-between;
+            padding-top: 15px;
         }
     }
 }
