+++ musasampo-frontend/src/components/profile-new-album/profile-new-album.styles.scss
@@ -19,13 +19,13 @@ $main-light-color: #393e46;
         flex-direction: column;
         margin: auto;
     
-        .title6 {
+        .title {
         font-size: 34px;
         margin-top: 0px;
         margin-bottom: 5px;
         }
     
-        .subtitle6 {
+        .subtitle {
         font-size: 26px;
         }
     }
@@ -48,9 +48,11 @@ $main-light-color: #393e46;
         margin-left: 50px;
         width: 50%;
 
-        .buttons6 {
-            float: right;
-            padding-top: 20px;
+        .buttons {
+            display: flex;
+            justify-content: space-between;
+            padding-top: 15px;
+            width: 100%
             }
     }
 }
