+++ musasampo-frontend/src/components/profile-new-band/profile-new-band.styles.scss
@@ -19,13 +19,13 @@ $main-light-color: #393e46;
         flex-direction: column;
         margin: auto;
     
-        .title5 {
+        .title {
         font-size: 34px;
         margin-top: 0px;
         margin-bottom: 5px;
         }
     
-        .subtitle5 {
+        .subtitle {
         font-size: 26px;
         }
     }
@@ -48,7 +48,7 @@ $main-light-color: #393e46;
         margin-left: 50px;
         width: 50%;
 
-        .buttons5 {
+        .buttons {
             float: right;
             padding-top: 25px;
             }
